@@ -26,21 +26,31 @@
     "check-docs": "cd docs && ./doc_checker.sh"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.4",
-    "@babel/plugin-proposal-object-rest-spread": "^7.4.4",
-    "@babel/plugin-transform-destructuring": "^7.4.4",
-    "@babel/polyfill": "^7.4.4",
-    "@babel/preset-env": "^7.4.4",
+    "@babel/cli": "^7.0.0",
+    "@babel/core": "^7.0.0",
+    "@babel/plugin-proposal-class-properties": "^7.0.0",
+    "@babel/plugin-proposal-decorators": "^7.0.0",
+    "@babel/plugin-proposal-export-namespace-from": "^7.0.0",
+    "@babel/plugin-proposal-function-sent": "^7.0.0",
+    "@babel/plugin-proposal-json-strings": "^7.0.0",
+    "@babel/plugin-proposal-numeric-separator": "^7.0.0",
+    "@babel/plugin-proposal-throw-expressions": "^7.0.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
+    "@babel/plugin-syntax-import-meta": "^7.0.0",
+    "@babel/polyfill": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
     "@babel/preset-react": "^7.0.0",
-    "@mapbox/mapbox-gl-draw": "^1.1.1",
+    "@mapbox/mapbox-gl-draw": "^1.0.4",
     "@mapbox/mapbox-gl-draw-static-mode": "^1.0.1",
     "@turf/area": "^5.0.4",
     "@turf/distance": "^5.0.4",
-    "babel-cli": "^6.26.0",
-    "babel-eslint": "^7.2.3",
-    "babel-jest": "^24.8.0",
-    "babel-loader": "8.0.5",
-    "babel-plugin-module-resolver": "^3.2.0",
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-eslint": "^9.0.0",
+    "babel-jest": "^23.4.2",
+    "babel-loader": "^8.0.6",
+    "babel-plugin-dynamic-import-node": "^2.3.0",
+    "babel-plugin-module-resolver": "^3.1.0",
+    "babel-plugin-react-transform": "^3.0.0",
     "canvas": "^1.6.8",
     "coveralls": "^2.13.1",
     "css-loader": "0.28.7",
@@ -57,7 +67,7 @@
     "eslint-plugin-jsx-a11y": "^5.0.2",
     "eslint-plugin-promise": "^3.6.0",
     "eslint-plugin-react": "^7.1.0",
-    "extract-text-webpack-plugin": "3.0.2",
+    "extract-text-webpack-plugin": "^4.0.0-beta.0",
     "fs-extra": "^4.0.2",
     "handlebars": "^4.0.10",
     "jest": "^20.0.4",
@@ -71,7 +81,7 @@
     "metalsmith-rootpath": "^1.0.4",
     "mixin": "^0.2.0",
     "nock": "^9.0.13",
-    "node-sass": "^4.12",
+    "node-sass": "^4.5.3",
     "ol": "^5.0.0",
     "ol-mapbox-style": "^3.0.1",
     "raf": "^3.4.0",
@@ -85,30 +95,35 @@
     "redux-mock-store": "^1.2.3",
     "sass-loader": "6.0.6",
     "style-loader": "0.19.0",
-    "uglifyjs-webpack-plugin": "1.0.1",
-    "uuid": "^3.3.2",
-    "webpack": "^3.8.1",
-    "webpack-dev-server": "2.9.3",
+    "uglifyjs-webpack-plugin": "^2.1.3",
+    "webpack": "^4.35.0",
+    "webpack-cli": "^3.3.5",
+    "webpack-dev-server": "^3.7.2",
     "xmldom": "^0.1.27"
   },
   "dependencies": {
     "@mapbox/mapbox-gl-style-spec": "^12.0.0",
     "fetch-jsonp": "^1.1.3",
     "isomorphic-fetch": "^2.2.1",
-    "prop-types": "^15.7.2",
+    "prop-types": "^15.5.10",
     "react-docgen": "^2.19.0",
-    "redux-saga": "^0.16.0"
+    "redux-saga": "^0.16.0",
+    "uuid": "^3.1.0"
   },
   "peerDependencies": {
     "@mapbox/mapbox-gl-draw": "^1.0.4",
     "@mapbox/mapbox-gl-draw-static-mode": "^1.0.1",
     "@turf/area": "^5.0.4",
     "@turf/distance": "^5.0.4",
+    "mapbox-gl": "^0.42.0",
+    "ol": "^5.0.0",
     "ol-mapbox-style": "^3.0.1",
     "react": "^16.0.0",
     "react-dnd": "^2.5.4",
     "react-dnd-html5-backend": "^2.5.4",
-    "react-dom": "^16.0.0"
+    "react-dom": "^16.0.0",
+    "react-redux": "^5.0.5",
+    "redux": "^3.6.0"
   },
   "repository": {
     "type": "git",
