@@ -1,11 +1,11 @@
-const { lstatSync, readdirSync } = require('fs')
+const {lstatSync, readdirSync} = require('fs');
 const path = require('path');
-const { join } = require('path');
+const {join} = require('path');
 const conf = require('./tasks/config');
 
-const isDirectory = source => lstatSync(source).isDirectory()
+const isDirectory = source => lstatSync(source).isDirectory();
 const getDirectories = source =>
-  readdirSync(source).map(name => join(source, name)).filter(isDirectory)
+  readdirSync(source).map(name => join(source, name)).filter(isDirectory);
 
 module.exports = {
   getEntries(dev) {
@@ -16,8 +16,8 @@ module.exports = {
       if (name !== 'resources' && (dev || conf.skip.indexOf(name) === -1)) {
         entry[name] = dev ? [
           'webpack/hot/only-dev-server',
-          'babel-polyfill'
-        ] : ['babel-polyfill'];
+          '@babel/polyfill'
+        ] : ['@babel/polyfill'];
         entry[name].push(`.${path.sep}${subDirs[i]}${path.sep}app.js`);
       }
     }
